Restructure server actions
